Changed to support float values

+++ ThingSpeakLinux.h
@@ -12,4 +12,4 @@ Date:   05/2016
 #define SEND_OK                      4
 
 //global prototypes
-char SendDataToThingSpeak(int FieldNo, int * FieldArray, char * Key, int SizeOfKey);
+char SendDataToThingSpeak(int FieldNo, float * FieldArray, char * Key, int SizeOfKey);
